Sort options
Sorting options alphabeticaly in getopts' optstring and the while loop is easier to follow and helps against double declaration and missing options.

While here remove the first s that is messing with the leading colon.
In some ksh implementation a double colon means the argument is
optional, which is probably not what was desired here. Unused options
have also been removed from the optstring.